+++ Diary/ViewController/DetailViewController.swift
@@ -2,7 +2,7 @@
 //  DetailViewController.swift
 //  Diary
 //
-//  Created by yyss99, Jusbug on 2023/08/29.
+//  Created by 박종화 on 2023/08/30.
 //
 
 import UIKit
@@ -10,68 +10,24 @@ import UIKit
 class DetailViewController: UIViewController {
     @IBOutlet weak var titleTextView: UITextView!
     @IBOutlet weak var bodyTextView: UITextView!
-    let placeHolderText = "Input Text"
+    private var sample: Sample?
+    
+    init?(sample: Sample, coder: NSCoder) {
+        self.sample = sample
+        super.init(coder: coder)
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        initTitleTextView()
-        initBodyTextView()
-        configureNavigationTitle()
-        
+        configureTextView()
     }
     
-    override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
-        self.titleTextView.resignFirstResponder()
-    }
-    
-    private func initTitleTextView() {
-        titleTextView.text = placeHolderText
-        titleTextView.textColor = .lightGray
-        titleTextView.layer.borderWidth = 1
-        titleTextView.delegate = self
-    }
-    
-    private func initBodyTextView() {
-        bodyTextView.text = placeHolderText
-        bodyTextView.textColor = .lightGray
-        bodyTextView.layer.borderWidth = 1
-        bodyTextView.delegate = self
-    }
-    
-    private func configureNavigationTitle() {
-        let dateFormatter = DateFormatter()
-        dateFormatter.locale = Locale(identifier: "ko_KR")
-        dateFormatter.dateFormat = "yyyy년 MM월 dd일"
-        
-        let today = Date()
-        let formattedTodayDate = dateFormatter.string(from: today)
-        
-        self.navigationItem.title = formattedTodayDate
-    }
-}
-
-extension DetailViewController: UITextViewDelegate {
-    func textViewDidBeginEditing(_ textView: UITextView) {
-        if titleTextView.text == placeHolderText {
-            titleTextView.text = nil
-            titleTextView.textColor = .black
-        }
-        
-        if bodyTextView.text == placeHolderText {
-            bodyTextView.text = nil
-            bodyTextView.textColor = .black
-        }
-    }
-    
-    func textViewDidEndEditing(_ textView: UITextView) {
-        if titleTextView.text.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
-            titleTextView.text = placeHolderText
-            titleTextView.textColor = .lightGray
-        }
-        
-        if bodyTextView.text.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
-            bodyTextView.text = placeHolderText
-            bodyTextView.textColor = .lightGray
-        }
+    private func configureTextView() {
+        titleTextView.text = sample?.title
+        bodyTextView.text = sample?.body
     }
 }
